refactor(terminal): extract animation components into dedicated module
- Move AnimatedSpan and TypingAnimation from Terminal.tsx to AnimatedComponents.tsx
- Lazy-load SpotifyPlayer with next/dynamic (ssr: false) in Terminal shell
- Update all command render imports to use new AnimatedComponents module
- Remove unused React imports from CommandBash, CEducation, CExperiences, CNowPlaying
- Slim down Terminal.tsx to a pure layout shell component
