+++ components/Terminal.tsx
@@ -1,98 +1,13 @@
 "use client";
 
 import { cn } from "@/lib/utils";
-import { motion, MotionProps } from "motion/react";
-import { useEffect, useRef, useState } from "react";
+import dynamic from "next/dynamic";
 import TerminalInput from "./TerminalInput";
 import { ModeToggle } from "./ModeToggle";
-import SpotifyPlayer from "./SpotifyPlayer";
 
-interface AnimatedSpanProps extends MotionProps {
-  children: React.ReactNode;
-  delay?: number;
-  className?: string;
-}
-
-export const AnimatedSpan = ({
-  children,
-  delay = 0,
-  className,
-  ...props
-}: AnimatedSpanProps) => (
-  <motion.div
-    initial={{ opacity: 0, y: -5 }}
-    animate={{ opacity: 1, y: 0 }}
-    transition={{ duration: 0.3, delay: delay / 1000 }}
-    className={cn("grid text-xs font-normal tracking-tight", className)}
-    {...props}
-  >
-    {children}
-  </motion.div>
-);
-
-interface TypingAnimationProps extends MotionProps {
-  children: string;
-  className?: string;
-  duration?: number;
-  delay?: number;
-  as?: React.ElementType;
-}
-
-export const TypingAnimation = ({
-  children,
-  className,
-  duration = 60,
-  delay = 0,
-  as: Component = "span",
-  ...props
-}: TypingAnimationProps) => {
-  if (typeof children !== "string") {
-    throw new Error("TypingAnimation: children must be a string. Received:");
-  }
-
-  const MotionComponent = motion.create(Component, {
-    forwardMotionProps: true,
-  });
-
-  const [displayedText, setDisplayedText] = useState<string>("");
-  const [started, setStarted] = useState(false);
-  const elementRef = useRef<HTMLElement | null>(null);
-
-  useEffect(() => {
-    const startTimeout = setTimeout(() => {
-      setStarted(true);
-    }, delay);
-    return () => clearTimeout(startTimeout);
-  }, [delay]);
-
-  useEffect(() => {
-    if (!started) return;
-
-    let i = 0;
-    const typingEffect = setInterval(() => {
-      if (i < children.length) {
-        setDisplayedText(children.substring(0, i + 1));
-        i++;
-      } else {
-        clearInterval(typingEffect);
-      }
-    }, duration);
-
-    return () => {
-      clearInterval(typingEffect);
-    };
-  }, [children, duration, started]);
-
-  return (
-    <MotionComponent
-      ref={elementRef}
-      className={cn("text-sm font-normal tracking-tight", className)}
-      {...props}
-    >
-      {displayedText}
-    </MotionComponent>
-  );
-};
+const SpotifyPlayer = dynamic(() => import("./SpotifyPlayer"), {
+  ssr: false,
+});
 
 interface TerminalProps {
   children: React.ReactNode;
@@ -103,8 +18,8 @@ export const Terminal = ({ children, className }: TerminalProps) => {
   return (
     <div
       className={cn(
-        "z-0 h-full flex flex-col max-h-full overflow-hidden md:max-h-[500px] w-full rounded-xl border border-border bg-background",
-        className
+        "z-0 h-full flex flex-col max-h-full overflow-hidden md:h-[500px] md:max-h-[500px] w-full rounded-xl border border-border bg-background",
+        className,
       )}
     >
       <div className="flex relative border-b border-border p-4">
