+++ components/commands/renders/spotify/CNowPlaying.tsx
@@ -1,8 +1,7 @@
 import { getNowPlaying } from "@/app/actions";
-import { AnimatedSpan } from "@/components/Terminal";
+import { AnimatedSpan } from "@/components/AnimatedComponents";
 import Image from "next/image";
 import Link from "next/link";
-import React from "react";
 import SpotifyQuery from "./SpotifyQuery";
 
 function CNowPlaying() {
